People controller separated by plurality
