+++ WebApi.Client/MainWindow.xaml.cs
@@ -45,7 +45,7 @@ namespace WebApi.Client
             using (var client =  new HttpClient())
             {
                 var content = new StringContent(JsonConvert.SerializeObject(PrimePerson()),Encoding.UTF8, "application/json");
-                var response = await client.PostAsync(Tie + "People", content);
+                var response = await client.PostAsync(Tie + "Person", content);
 
                 if (response.IsSuccessStatusCode)
                 {
@@ -60,7 +60,7 @@ namespace WebApi.Client
         {
             using (var client = new HttpClient())
             {
-                var response = await client.GetAsync(new Uri(Tie + $"People/byName{PersonNameTextBox.Text}"));
+                var response = await client.GetAsync(new Uri(Tie + $"Person/byName{PersonNameTextBox.Text}"));
 
                 if(response.IsSuccessStatusCode)
                 {
@@ -76,7 +76,7 @@ namespace WebApi.Client
             using (var client = new HttpClient())
             {
                 var content = new StringContent(JsonConvert.SerializeObject(PrimePerson()), Encoding.UTF8, "application/json");
-                var response = await client.PutAsync(Tie + "People", content);
+                var response = await client.PutAsync(Tie + "Person", content);
 
                 if (response.IsSuccessStatusCode)
                 {
